refactor(story-engine): flatten internal api v2 package
